refactor: rename lib directory to utils
