+++ src/Session.ts
@@ -9,26 +9,26 @@ import type {Connection} from "./adapters/Connection";
 import {FetchConnection} from "./adapters/FetchConnection";
 import {NodeHttp1Connection} from "./adapters/NodeHttp1Connection";
 import {NodeHttp2CompatConnection} from "./adapters/NodeHttp2CompatConnection";
-import {SseError} from "./lib/SseError";
-import {type EventMap, TypedEmitter} from "./lib/TypedEmitter";
-import {applyHeaders} from "./lib/applyHeaders";
+import {SseError} from "./utils/SseError";
+import {type EventMap, TypedEmitter} from "./utils/TypedEmitter";
+import {applyHeaders} from "./utils/applyHeaders";
 import {
 	type PushFromIterable,
 	createPushFromIterable,
-} from "./lib/createPushFromIterable";
+} from "./utils/createPushFromIterable";
 import {
 	type PushFromStream,
 	createPushFromStream,
-} from "./lib/createPushFromStream";
-import {generateId} from "./lib/generateId";
+} from "./utils/createPushFromStream";
+import {generateId} from "./utils/generateId";
 import {
 	type SanitizerFunction,
 	sanitize as defaultSanitizer,
-} from "./lib/sanitize";
+} from "./utils/sanitize";
 import {
 	type SerializerFunction,
 	serialize as defaultSerializer,
-} from "./lib/serialize";
+} from "./utils/serialize";
 
 interface SessionOptions<State = DefaultSessionState>
 	extends Pick<EventBufferOptions, "serializer" | "sanitizer"> {
