+++ src/utils/SseError.ts
